healthcabinet/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.84.0",
    "boto3>=1.34",
    "arq>=0.27.0",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "langchain-anthropic>=1.0.1",
    "langchain-core>=1.0.4",
    "langgraph>=1.1.3",
    "langsmith>=0.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic-settings>=2.0",
    "PyJWT>=2.10.0",
    "python-multipart>=0.0.22",
    "redis>=5.3.1",
    "sentry-sdk[fastapi]>=2.54.0",
    "sqlalchemy[asyncio]>=2.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning:passlib"]
markers = ["integration: marks tests that make real external API calls (deselect with '-m \"not integration\"')"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]  # B008: FastAPI requires Depends() in function signatures

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
